Call Management Service
This system allows the users to communicate directly with the company’s Customer Care Department.
This allows the users to feel free and comfortable to interact with them and share their problems with the employees.
The users have a 24*7 Round About Facility to have an interaction anytime.
Continuous Feedback of the users is taken and on the basis of that a Monthly Complaint Report is generated which allows the Product Quality to be updated and improved.